fix: 읽지 않는 알림 갯수 일치 않는 문제 (#362)
- 원인 : TxListener 에서 로직 순서로 인한 문제
- 해결 : 알림 저장 로직과 이벤트 송신 로직 위치 변경
