+++ src/main/java/kr/mywork/domain/post/listener/PostNotificationTxListener.java
@@ -1,4 +1,4 @@
-package kr.mywork.domain.notification.listener.event;
+package kr.mywork.domain.post.listener;
 
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
@@ -7,29 +7,45 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 
+import kr.mywork.domain.post.listener.event.ReviewNotificationCreateEvent;
 import kr.mywork.domain.notification.service.NotificationService;
 import kr.mywork.domain.notification.service.RealTimeNotificationService;
+import kr.mywork.domain.post.listener.event.PostApprovalNotificationEvent;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 
-@Slf4j
 @Component
 @RequiredArgsConstructor
-public class NotificationTxEventListener {
+public class PostNotificationTxListener {
 
 	private final RealTimeNotificationService realTimeNotificationService;
 	private final NotificationService notificationService;
 
+	@Transactional(propagation = Propagation.REQUIRES_NEW)
+	@Async("eventTaskExecutor")
+	@TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+	public void handlePostApprovalAlarmEvent(final PostApprovalNotificationEvent event) {
+		saveApprovalNotification(event);
+		final long unreadCount = notificationService.countUnreadNotifications(event.authorId());
+		realTimeNotificationService.sendNotification(event.authorId(), "notification-unread-count", unreadCount);
+	}
+
+	private void saveApprovalNotification(final PostApprovalNotificationEvent event) {
+		notificationService.save(
+			event.authorId(), event.authorName(), event.postTitle(), event.actorName(), event.actorId(),
+			event.targetType(), event.postId(), event.notificationActionType(), event.modifiedAt(), event.projectId(),
+			event.projectStepId());
+	}
+
 	@Async(value = "eventTaskExecutor")
 	@Transactional(propagation = Propagation.REQUIRES_NEW)
 	@TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-	public void handleNotificationCreateEvent(final NotificationCreateEvent event) {
+	public void handleNotificationCreateEvent(final ReviewNotificationCreateEvent event) {
+		saveReviewNotification(event);
 		long unreadCount = notificationService.countUnreadNotifications(event.authorId());
 		realTimeNotificationService.sendNotification(event.authorId(), "notification-unread-count", unreadCount);
-		saveNotification(event);
 	}
 
-	private void saveNotification(final NotificationCreateEvent event) {
+	private void saveReviewNotification(final ReviewNotificationCreateEvent event) {
 		notificationService.save(
 			event.authorId(),
 			event.authorName(),
